move them to ChatConfig
